+++ packages/renderer/src/components/Startup.vue
@@ -2,7 +2,7 @@
   <div class="vtj-startup__wrapper">
     <div class="vtj-startup">
       <h1 class="vtj-startup__name">
-        <span>{{ props.name }}</span>
+        <span class="span">{{ props.name }}</span>
       </h1>
       <div class="vtj-startup__tagline">{{ props.tagline }}</div>
       <div class="vtj-startup__actions">
@@ -53,6 +53,7 @@
     grid-row: 1;
     width: 100%;
     max-width: 800px;
+    margin: 0 10px;
 
     &__wrapper {
       height: 100%;
@@ -117,6 +118,7 @@
         cursor: pointer;
         display: inline-block;
         line-height: 40px;
+        text-align: center;
 
         @media (max-width: 768px) {
           width: 100%;