+++ packages/uni/src/setup/app.ts
@@ -2,7 +2,7 @@ import type { UniConfig, JSFunction } from '@vtj/core';
 import type { Provider } from '@vtj/renderer';
 import type { SetupUniAppOptions, UniRoute } from '../types';
 import { mergeOptions } from '../utils';
-import { APP_LIFE_CYCLE } from '../constants';
+import { APP_LIFE_CYCLE, ROUTE_PAGE_BASE_PATH } from '../constants';
 import {
   injectUniConfig,
   injectUniGlobal,
@@ -52,16 +52,15 @@ export function createUniAppComponent(
 
 export async function createUniRoutes(
   provider: Provider,
-  createRenderer: any,
   includeBlock: boolean = false,
-  basePath: string = '/pages'
+  basePath: string = ROUTE_PAGE_BASE_PATH
 ) {
   const pages = provider.project?.pages || [];
   const routes: UniRoute[] = [];
   for (const page of pages) {
     const dsl = await provider.getDsl(page.id);
     if (!dsl) continue;
-    const { renderer } = createRenderer({ dsl, components: {} });
+    const { renderer } = provider.createDslRenderer(dsl);
     const home = provider.project?.homepage === page.id;
     routes.push({
       id: page.id,
@@ -78,13 +77,10 @@ export async function createUniRoutes(
 
   const homeRoute = routes.find((route) => !!route.home) || routes[0];
   if (homeRoute) {
+    const homePath = basePath === ROUTE_PAGE_BASE_PATH ? '/' : basePath;
     routes.unshift({
       ...homeRoute,
-      path: '/'
-    });
-    routes.unshift({
-      ...homeRoute,
-      path: basePath
+      path: homePath
     });
   }
 
@@ -93,7 +89,7 @@ export async function createUniRoutes(
     for (const block of blocks) {
       const dsl = await provider.getDsl(block.id);
       if (!dsl) continue;
-      const { renderer } = createRenderer({ dsl, components: {} });
+      const { renderer } = provider.createDslRenderer(dsl);
       routes.push({
         id: block.id,
         path: `${basePath}/${block.id}`,